Revert "Splitting area selection and layer selection in two components"
This reverts commit 99493515df5ccbd7b3014706cdb405d2c1dede36.
