+++ frontend/src/public/global.css
@@ -12,8 +12,8 @@ body {
 	font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen-Sans, Ubuntu, Cantarell, "Helvetica Neue", sans-serif;
 }
 
-/*
 a {
+	/*color: rgb(0,100,200);*/
 	color:#4d88c7;
 	text-decoration: none;
 }
@@ -25,7 +25,6 @@ a:hover {
 a:visited {
 	color: rgb(0,80,160);
 }
-*/
 
 label {
 	display: block;
