+++ frontend/src/src/App.svelte
@@ -5,7 +5,6 @@
 </script>
 
 <style>
-  
 :global(body) {
 	width: 100%;
 	height: 100%;
@@ -20,15 +19,11 @@
   background-size: cover;
   background: url('/images/clear-icon.png');
 }
-
-
 #main {
 	width: 100%;
 	height: 100%;
 	display: flex;
 }
-
-
 :global(.close_button) {
   right: 0;
   position: absolute;
